+++ tests/unit/tooltip/deprecated.js
@@ -1,6 +1,6 @@
 define( [
 	"jquery",
-	"ui/tooltip"
+	"ui/widgets/tooltip"
 ], function( $ ) {
 
 module( "tooltip: (deprecated) options" );
